[#5] MVVM+C 분리
