+++ Hyodor/Hyodor/Modules/Alert/AlertView.swift
@@ -1,0 +1,17 @@
+//
+//  AlertView.swift
+//  Hyodor
+//
+//  Created by 김상준 on 4/28/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct AlertView: View {
+    var body: some View {
+        Text("Hello, World!")
+    }
+
+
+}
