+++ Hyodor/Hyodor/Modules/Home/Home/Component/SectionHeader.swift
@@ -1,0 +1,17 @@
+//
+//  SectionHeader.swift
+//  Hyodor
+//
+//  Created by 김상준 on 4/28/25.
+//
+
+import SwiftUI
+
+struct SectionHeader: View {
+    let title: String
+    var body: some View {
+        Text(title)
+            .font(.headline)
+            .padding(.bottom, 4)
+    }
+}
